Initial Upload for a Sequencer for Akai MPX16
It uses a Arduino UNO, Softserial to Pin 2/3 and 3 PCF8574 Multiplexers. 2 x PCF8574 to manage 16 LEDs and 16 Buttons and the third PCF7574 to manage a LCD-Display 16x2.  I have made the initial Code around 2015 and tried the reuse the electronics for a newer AKAI MPX16. There are no sounds created by the Arduino UNO. All sounds are created by the Akai MPX and the UNO is only creating the MIDI-Signals. Video: https://www.youtube.com/watch?v=Tr9Y971fDvM